Refactor: Align wordBreak folder structure with 3sum
Moved the wordBreak problem files (`wordBreak.ts`, `wordBreak.test.ts`) into a dedicated directory `packages/backend/src/problem/free/wordBreak/`.

Created standard placeholder files (`code/index.ts`, `code/solution.ts`, `groups.ts`, `problem.ts`, `steps.ts`, `testcase.ts`, `types.ts`, `variables.ts`) within the new directory to match the convention used by other problems like 3sum.

Renamed `wordBreak.test.ts` to `index.test.ts`.

Updated relative import paths within the moved files to reflect the new directory depth.
