+++ packages/backend/src/problem/free/wordBreak/index.test.ts
@@ -1,6 +1,6 @@
 import { it } from "bun:test";
 import { problem } from "./wordBreak"; // Import from the sibling .ts file
-import { runTests } from "../core/test"; // Adjust path relative to the new test file
+import { runTests } from "../../core/test"; // Adjust path relative to the new test file
 
 it(problem.id, () => {
   runTests(problem);
